moved flash waitstates to lpc17xx
