+++ targets/lpc1756/io/system.hpp
@@ -4,11 +4,16 @@
 #include <lpc1756.hpp>
 
 #include <targets/core/nxp/lpc175x/system.hpp>
+#include <targets/core/nxp/lpc17xx/system.hpp>
 
 namespace klib::lpc1756::io::system {
     // import all the static classes from the core
     // library in the current system namespace
     using namespace core::lpc175x::io::system;
+
+    // import the generic lpc17xx system drivers as 
+    // well
+    using namespace core::lpc17xx::io::system;
 }
 
 #endif
